refactor: Mission
