+++ isol/server-common/src/main/java/shinhan/server_common/domain/mission/dto/MissionFindOneResponse.java
@@ -1,4 +1,4 @@
-package shinhan.server_parent.domain.mission.dto;
+package shinhan.server_common.domain.mission.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
