+++ isol/server-common/src/main/java/shinhan/server_common/domain/mission/dto/MissionSaveRequest.java
@@ -1,4 +1,4 @@
-package shinhan.server_parent.domain.mission.dto;
+package shinhan.server_common.domain.mission.dto;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
